% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-1495_abstraction.R
\name{count_loss}
\alias{count_loss}
\title{Count Patients with Specific Marker Loss Patterns}
\usage{
count_loss(data, loss_levels, must_have_loss, must_not_have_loss)
}
\arguments{
\item{data}{A data frame containing marker columns and patient IDs.}

\item{loss_levels}{A vector of values considered as "loss" for markers.}

\item{must_have_loss}{A character vector of marker column names that should not show loss.}

\item{must_not_have_loss}{A character vector of marker column names that should not not show loss.}
}
\value{
An integer count of distinct patients matching the criteria.
}
\description{
Count Patients with Specific Marker Loss Patterns
}
