% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-1495_abstraction.R
\name{cox_uv_regression}
\alias{cox_uv_regression}
\title{Generate a Uni variate Cox Regression Table}
\usage{
cox_uv_regression(data, outcome_prefix, covariates, filter_expr = NULL)
}
\arguments{
\item{data}{A data frame containing survival and covariate variables.}

\item{outcome_prefix}{A string like "os", "dss", or "pfs".}

\item{covariates}{Character vector of covariate column names.}

\item{filter_expr}{Optional filtering expression (e.g., mmr_loss_type_2 == "MLH1 or PMS2/MLH1 loss").}
}
\value{
A gtsummary table object.
}
\description{
Generate a Uni variate Cox Regression Table
}
