% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-1495_abstraction.R
\name{cox_uv_regression_by_group}
\alias{cox_uv_regression_by_group}
\title{Generate Cox Regression Tables Grouped by a Variable}
\usage{
cox_uv_regression_by_group(data, outcome_prefix, covariates, group_var)
}
\arguments{
\item{data}{A data frame.}

\item{outcome_prefix}{A string like "os", "dss", or "pfs".}

\item{covariates}{Character vector of covariate column names.}

\item{group_var}{Name of the grouping variable (e.g. "stage").}
}
\value{
A gtsummary stratified table object.
}
\description{
Generate Cox Regression Tables Grouped by a Variable
}
