% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-1495_abstraction.R
\name{create_bar_plot}
\alias{create_bar_plot}
\title{Create a Bar Plot for a Categorical Variable}
\usage{
create_bar_plot(data, x_var, plot_title)
}
\arguments{
\item{data}{A data frame containing the variable to plot.}

\item{x_var}{A string name of the categorical variable to plot on the x-axis.}

\item{plot_title}{A string for the plot title and x-axis label.}
}
\value{
A ggplot2 bar plot object.
}
\description{
Create a Bar Plot for a Categorical Variable
}
