% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-1495_abstraction.R
\name{plot_km_curves}
\alias{plot_km_curves}
\title{Generate and Print Kaplan-Meier Curves}
\usage{
plot_km_curves(args, caps)
}
\arguments{
\item{args}{A named list of arguments to pass to `doKMPlots`.}

\item{caps}{A string title to use for the plot.}
}
\value{
A printed ggplot object with KM curves.
}
\description{
Generate and Print Kaplan-Meier Curves
}
